+++ rundeck/error.go
@@ -1,0 +1,7 @@
+package rundeck
+
+type NotFoundError struct{}
+
+func (err NotFoundError) Error() string {
+	return "not found"
+}