+++ rundeck/resource_public_key_test.go
@@ -1,18 +1,20 @@
 package rundeck
 
 import (
+	"context"
 	"fmt"
+	"io/ioutil"
 	"strings"
 	"testing"
 
-	"github.com/apparentlymart/go-rundeck-api/rundeck"
+	"github.com/rundeck/go-rundeck/rundeck"
 
 	"github.com/hashicorp/terraform/helper/resource"
 	"github.com/hashicorp/terraform/terraform"
 )
 
 func TestAccPublicKey_basic(t *testing.T) {
-	var key rundeck.KeyMeta
+	var key rundeck.StorageKeyListResponse
 	var keyMaterial string
 
 	resource.Test(t, resource.TestCase{
@@ -25,17 +27,17 @@ func TestAccPublicKey_basic(t *testing.T) {
 				Check: resource.ComposeTestCheckFunc(
 					testAccPublicKeyCheckExists("rundeck_public_key.test", &key, &keyMaterial),
 					func(s *terraform.State) error {
-						if expected := "keys/terraform_acceptance_tests/public_key"; key.Path != expected {
+						if expected := "keys/terraform_acceptance_tests/public_key"; *key.Path != expected {
 							return fmt.Errorf("wrong path; expected %v, got %v", expected, key.Path)
 						}
-						if !strings.HasSuffix(key.URL, "/storage/keys/terraform_acceptance_tests/public_key") {
+						if !strings.HasSuffix(*key.URL, "/storage/keys/terraform_acceptance_tests/public_key") {
 							return fmt.Errorf("wrong URL; expected to end with the key path")
 						}
-						if expected := "file"; key.ResourceType != expected {
-							return fmt.Errorf("wrong resource type; expected %v, got %v", expected, key.ResourceType)
+						if expected := "file"; *key.Type != expected {
+							return fmt.Errorf("wrong resource type; expected %v, got %v", expected, *key.Type)
 						}
-						if expected := "public"; key.KeyType != expected {
-							return fmt.Errorf("wrong key type; expected %v, got %v", expected, key.KeyType)
+						if expected := rundeck.Public; key.Meta.RundeckKeyType != expected {
+							return fmt.Errorf("wrong key type; expected %v, got %v", expected, key.Meta.RundeckKeyType)
 						}
 						if !strings.Contains(keyMaterial, "test+public+key+for+terraform") {
 							return fmt.Errorf("wrong key material")
@@ -48,14 +50,17 @@ func TestAccPublicKey_basic(t *testing.T) {
 	})
 }
 
-func testAccPublicKeyCheckDestroy(key *rundeck.KeyMeta) resource.TestCheckFunc {
+func testAccPublicKeyCheckDestroy(key *rundeck.StorageKeyListResponse) resource.TestCheckFunc {
 	return func(s *terraform.State) error {
-		client := testAccProvider.Meta().(*rundeck.Client)
-		_, err := client.GetKeyMeta(key.Path)
-		if err == nil {
+		client := testAccProvider.Meta().(*rundeck.BaseClient)
+		ctx := context.Background()
+
+		resp, err := client.StorageKeyGetMetadata(ctx, *key.Path)
+
+		if resp.StatusCode == 200 {
 			return fmt.Errorf("key still exists")
 		}
-		if _, ok := err.(*rundeck.NotFoundError); !ok {
+		if resp.StatusCode != 404 {
 			return fmt.Errorf("got something other than NotFoundError (%v) when getting key", err)
 		}
 
@@ -63,7 +68,7 @@ func testAccPublicKeyCheckDestroy(key *rundeck.KeyMeta) resource.TestCheckFunc {
 	}
 }
 
-func testAccPublicKeyCheckExists(rn string, key *rundeck.KeyMeta, keyMaterial *string) resource.TestCheckFunc {
+func testAccPublicKeyCheckExists(rn string, key *rundeck.StorageKeyListResponse, keyMaterial *string) resource.TestCheckFunc {
 	return func(s *terraform.State) error {
 		rs, ok := s.RootModule().Resources[rn]
 		if !ok {
@@ -74,18 +79,23 @@ func testAccPublicKeyCheckExists(rn string, key *rundeck.KeyMeta, keyMaterial *s
 			return fmt.Errorf("key id not set")
 		}
 
-		client := testAccProvider.Meta().(*rundeck.Client)
-		gotKey, err := client.GetKeyMeta(rs.Primary.ID)
-		if err != nil {
+		client := testAccProvider.Meta().(*rundeck.BaseClient)
+		ctx := context.Background()
+		gotKey, err := client.StorageKeyGetMetadata(ctx, rs.Primary.ID)
+		if gotKey.StatusCode == 404 || err != nil {
 			return fmt.Errorf("error getting key metadata: %s", err)
 		}
 
-		*key = *gotKey
+		*key = gotKey
 
-		*keyMaterial, err = client.GetKeyContent(rs.Primary.ID)
+		resp, err := client.StorageKeyGetMaterial(ctx, rs.Primary.ID)
 		if err != nil {
 			return fmt.Errorf("error getting key contents: %s", err)
 		}
+
+		respMaterial, err := ioutil.ReadAll(resp.Body)
+
+		*keyMaterial = string(respMaterial)
 
 		return nil
 	}